+++ src/main/java/com/seeyon/orgcenter/manage/function/entity/OrgFunction.java
@@ -27,10 +27,10 @@ public class OrgFunction implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value="ID",type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @TableField("APP_ID")
-    private Integer appId;
+    private Long appId;
 
     @TableField("NAME")
     private String name;
